change groupId to io.github.mahmoudi
