+++ src/main/java/io/github/mahmoudi/converter/DoubleTypeConverter.java
@@ -1,7 +1,7 @@
 /*
  * code https://github.com/jittagornp/excel-object-mapping
  */
-package com.mapper.excel.converter;
+package io.github.mahmoudi.converter;
 
 import java.math.BigDecimal;
 import org.slf4j.Logger;
