+++ src/main/java/io/github/mahmoudi/util/WorkbookCallback.java
@@ -1,7 +1,7 @@
 /*
  * code https://github.com/jittagornp/excel-object-mapping
  */
-package com.mapper.excel.util;
+package io.github.mahmoudi.util;
 
 import org.apache.poi.ss.usermodel.Workbook;
